refactor: change actions and selectors export
