+++ src/services/actions.ts
@@ -1,5 +1,3 @@
-export { default as userIsAuth } from './slice/auth/user-slice';
-export { default as userNotAuth } from './slice/auth/user-slice';
-export { default as setEmail } from './slice/auth/user-email-slice';
-export { default as resetEmail } from './slice/auth/user-email-slice';
-export { default as resetLoginError } from './slice/auth/auth-page-slice';
+export { userIsAuth, userNotAuth } from './slice/auth/user-slice';
+export { setEmail, resetEmail } from './slice/auth/user-email-slice';
+export { resetLoginError } from './slice/auth/auth-page-slice';
