+++ src/services/selectors.ts
@@ -1,4 +1,3 @@
-export { default as checkAuth } from './slice/auth/user-slice';
-export { default as getUserEmail } from './slice/auth/user-email-slice';
-export { default as getLoginError } from './slice/auth/auth-page-slice';
-export { default as getLoginErrorText } from './slice/auth/auth-page-slice';
+export { checkAuth } from './slice/auth/user-slice';
+export { getUserEmail } from './slice/auth/user-email-slice';
+export { getLoginError, getLoginErrorText } from './slice/auth/auth-page-slice';
